requirements.txt
python-telegram-bot==22.6
requests==2.32.3
python-dotenv==1.0.1
